+++ ui/src/admin/actions/influxdb.js
@@ -84,10 +84,6 @@ export const loadDatabases = databases => ({
   },
 })
 
-export const addRole = () => ({
-  type: 'INFLUXDB_ADD_ROLE',
-})
-
 export const addDatabase = () => ({
   type: 'INFLUXDB_ADD_DATABASE',
 })
@@ -129,14 +125,6 @@ export const syncRetentionPolicy = (database, stale, synced) => ({
     database,
     stale,
     synced,
-  },
-})
-
-export const editRole = (role, updates) => ({
-  type: 'INFLUXDB_EDIT_ROLE',
-  payload: {
-    role,
-    updates,
   },
 })
 
@@ -318,8 +306,6 @@ export const createRoleAsync = (url, role) => async dispatch => {
     dispatch(syncRole(role, data))
   } catch (error) {
     dispatch(errorThrown(error, notifyRoleCreationFailed(error.data.message)))
-    // undo optimistic update
-    setTimeout(() => dispatch(deleteRole(role)), REVERT_STATE_DELAY)
   }
 }