+++ ui/src/admin/components/influxdb/CreateRoleDialog.tsx
@@ -3,7 +3,6 @@ import {
   ComponentStatus,
   Form,
   Input,
-  InputType,
   OverlayBody,
   OverlayContainer,
   OverlayHeading,
@@ -11,54 +10,37 @@ import {
 } from 'src/reusable_ui'
 
 const minLen = 3
-export function validateUserName(name: string): boolean {
+export function validateRoleName(name: string): boolean {
   return name?.length >= minLen
 }
-export const validatePassword = validateUserName
 
 interface Props {
-  create: (user: {name: string; password: string}) => void
+  create: (role: {name: string}) => void
   setVisible: (visible: boolean) => void
   visible: boolean
 }
-const CreateUserDialog = ({visible, setVisible, create}: Props) => {
+const CreateRoleDialog = ({visible, setVisible, create}: Props) => {
   const [name, setName] = useState('')
-  const [password, setPassword] = useState('')
   const cancel = useCallback(() => {
     setName('')
-    setPassword('')
     setVisible(false)
   }, [])
   return (
     <OverlayTechnology visible={visible}>
       <OverlayContainer maxWidth={650}>
-        <OverlayHeading title="Create User" onDismiss={cancel} />
+        <OverlayHeading title="Create Role" onDismiss={cancel} />
         <OverlayBody>
           <Form>
-            <Form.Element label="User Name">
+            <Form.Element label="Role Name">
               <Input
                 value={name}
-                onChange={e => setName(e.target.value)}
                 autoFocus={true}
+                onChange={e => setName(e.target.value)}
                 status={
-                  validateUserName(name)
+                  validateRoleName(name)
                     ? ComponentStatus.Valid
                     : ComponentStatus.Default
                 }
-                testId="username--input"
-              />
-            </Form.Element>
-            <Form.Element label="Password">
-              <Input
-                value={password}
-                type={InputType.Password}
-                status={
-                  validatePassword(password)
-                    ? ComponentStatus.Valid
-                    : ComponentStatus.Default
-                }
-                onChange={e => setPassword(e.target.value)}
-                testId="password--input"
               />
             </Form.Element>
             <Form.Footer>
@@ -68,8 +50,8 @@ const CreateUserDialog = ({visible, setVisible, create}: Props) => {
                 </button>
                 <button
                   className="btn btn-sm btn-success"
-                  disabled={!(name && password)}
-                  onClick={() => create({name, password})}
+                  disabled={!name}
+                  onClick={() => create({name})}
                 >
                   Create
                 </button>
@@ -82,4 +64,4 @@ const CreateUserDialog = ({visible, setVisible, create}: Props) => {
   )
 }
 
-export default CreateUserDialog
+export default CreateRoleDialog