+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 1. [#5921](https://github.com/influxdata/chronograf/pull/5921): Manage InfluxDB users including their database permissions.
 1. [#5923](https://github.com/influxdata/chronograf/pull/5923): Manage InfluxDB roles including their database permissions.
 1. [#5925](https://github.com/influxdata/chronograf/pull/5925): Improve InfluxDB user creation.
+1. [#5926](https://github.com/influxdata/chronograf/pull/5926): Improve InfluxDB role creation.
 
 ### Bug Fixes